robot-server/requirements.txt
# Core dependencies
asyncio-mqtt==0.13.0
opencv-python==4.8.1.78
RPi.GPIO==0.7.1
websockets==11.0.3
numpy==1.24.3
Pillow==10.0.1

# Computer vision and AI
scikit-image==0.21.0
tensorflow-lite-runtime==2.13.0
imutils==0.5.4

# Communication and networking
requests==2.31.0
aiohttp==3.8.6

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Development and debugging
flake8==6.1.0
black==23.9.1

# Logging and monitoring
psutil==5.9.6

# Configuration management
python-dotenv==1.0.0

# Mathematical operations
scipy==1.11.4
